requirements.txt
mediapipe>=0.10.0
opencv-python>=4.8.0
pynput>=1.7.6
pyobjc-framework-Quartz>=9.0
